Move the UpgradeView to the views namespace.
